preprocess/dataset_generation/requirements.txt
tensorflow>=2.0
easydict
plyfile
numpy
scikit-image
pyyaml
h5py
opencv-python==4.8.1.78
tqdm
matplotlib
scipy
lpips
transforms3d
tensorboard==2.8.0
tensorboardX
torch
fairscale
fire
sentencepiece
protobuf==3.20.1 
proglog 
decorator 
msgpack 
simplejson 
click 
distro 
progress 
billiard==4.1.0
einops 
kornia
salesforce-lavis
gym==0.23.0
yacs==0.1.8